--- src/Functions/S2RectContains.cpp
@@ -1,108 +1,0 @@
-#if !defined(ARCADIA_BUILD)
-#    include "config_functions.h"
-#endif
-
-#if USE_S2_GEOMETRY
-
-#include <Columns/ColumnsNumber.h>
-#include <Columns/ColumnTuple.h>
-#include <DataTypes/DataTypesNumber.h>
-#include <DataTypes/DataTypeTuple.h>
-#include <Functions/FunctionFactory.h>
-#include <Common/typeid_cast.h>
-#include <common/range.h>
-
-#include "s2_fwd.h"
-
-class S2CellId;
-
-namespace DB
-{
-
-namespace ErrorCodes
-{
-    extern const int ILLEGAL_TYPE_OF_ARGUMENT;
-    extern const int NUMBER_OF_ARGUMENTS_DOESNT_MATCH;
-}
-
-namespace
-{
-
-/// TODO: Comment this
-class FunctionS2RectContains : public IFunction
-{
-public:
-    static constexpr auto name = "S2RectContains";
-
-    static FunctionPtr create(ContextPtr)
-    {
-        return std::make_shared<FunctionS2RectContains>();
-    }
-
-    std::string getName() const override
-    {
-        return name;
-    }
-
-    size_t getNumberOfArguments() const override { return 4; }
-
-    bool useDefaultImplementationForConstants() const override { return true; }
-
-    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override
-    {
-        for (size_t i = 0; i < getNumberOfArguments(); ++i)
-        {
-            const auto * arg = arguments[i].get();
-            if (!WhichDataType(arg).isUInt64()) {
-                throw Exception(
-                    ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT,
-                    "Illegal type {} of argument {} of function {}. Must be UInt64",
-                    arg->getName(), i, getName()
-                    );
-            }
-        }
-
-        return std::make_shared<DataTypeUInt8>();
-    }
-
-    ColumnPtr executeImpl(const ColumnsWithTypeAndName & arguments, const DataTypePtr &, size_t input_rows_count) const override
-    {
-        const auto * col_lo = arguments[0].column.get();
-        const auto * col_hi = arguments[1].column.get();
-        const auto * col_point = arguments[2].column.get();
-
-        auto dst = ColumnVector<UInt8>::create();
-        auto & dst_data = dst->getData();
-        dst_data.resize(input_rows_count);
-
-        for (const auto row : collections::range(0, input_rows_count))
-        {
-            const UInt64 lo = col_lo->getUInt(row);
-            const UInt64 hi = col_hi->getUInt(row);
-            const UInt64 point = col_point->getUInt(row);
-
-            S2CellId id_lo(lo);
-            S2CellId id_hi(hi);
-            S2CellId id_point(point);
-
-            S2LatLngRect rect(id_lo.ToLatLng(), id_hi.ToLatLng());
-
-            dst_data[row] = rect.Contains(id_point.ToLatLng());
-        }
-
-        return dst;
-    }
-
-};
-
-}
-
-void registerFunctionS2RectContains(FunctionFactory & factory)
-{
-    factory.registerFunction<FunctionS2RectContains>();
-}
-
-
-}
-
-#endif
