+++ src/Functions/s2CapContains.cpp
@@ -10,6 +10,7 @@
 #include <DataTypes/DataTypeTuple.h>
 #include <Functions/FunctionFactory.h>
 #include <Common/typeid_cast.h>
+#include <Common/NaNUtils.h>
 #include <common/range.h>
 
 #include "s2_fwd.h"
@@ -26,11 +27,20 @@ namespace ErrorCodes
 namespace
 {
 
-/// TODO: Comment this
+/**
+ * The cap represents a portion of the sphere that has been cut off by a plane.
+ * It is defined by a point on a sphere and a radius in degrees.
+ * Imagine that we draw a line through the center of the sphere and our point.
+ * An infinite number of planes pass through this line, but any plane will intersect the cap in two points.
+ * Thus the angle is defined by one of this points and the entire line.
+ * So, the radius of Pi/2 defines a hemisphere and the radius of Pi defines a whole sphere.
+ *
+ * This function returns whether a cap contains a point.
+ */
 class FunctionS2CapContains : public IFunction
 {
 public:
-    static constexpr auto name = "S2CapContains";
+    static constexpr auto name = "s2CapContains";
 
     static FunctionPtr create(ContextPtr)
     {
@@ -52,6 +62,7 @@ public:
         {
             const auto * arg = arguments[index].get();
 
+            /// Radius
             if (index == 1)
             {
                 if (!WhichDataType(arg).isFloat64())
@@ -64,8 +75,7 @@ public:
                 throw Exception(
                     ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT,
                     "Illegal type {} of argument {} of function {}. Must be UInt64",
-                    arg->getName(), index + 1, getName()
-                    );
+                    arg->getName(), index + 1, getName());
             }
         }
 
@@ -78,25 +88,24 @@ public:
         const auto * col_degrees = arguments[1].column.get();
         const auto * col_point = arguments[2].column.get();
 
-        auto dst = ColumnVector<UInt8>::create();
+        auto dst = ColumnUInt8::create();
         auto & dst_data = dst->getData();
-        dst_data.resize(input_rows_count);
+        dst_data.reserve(input_rows_count);
 
         for (const auto row : collections::range(0, input_rows_count))
         {
             const UInt64 center = col_center->getUInt(row);
             const Float64 degrees = col_degrees->getFloat64(row);
-            const UInt64 point = col_point->getInt(row);
+            const UInt64 point = col_point->getUInt(row);
 
             S1Angle angle = S1Angle::Degrees(degrees);
             S2Cap cap(S2CellId(center).ToPoint(), angle);
 
-            dst_data[row] = cap.Contains(S2CellId(point).ToPoint());
+            dst_data.emplace_back(cap.Contains(S2CellId(point).ToPoint()));
         }
 
         return dst;
     }
-
 };
 
 }
