+++ src/Functions/s2RectAdd.cpp
@@ -28,15 +28,15 @@ namespace ErrorCodes
 namespace
 {
 
-/// TODO: Comment this
-class FunctionS2RectIntersection : public IFunction
+
+class FunctionS2RectAdd : public IFunction
 {
 public:
-    static constexpr auto name = "S2RectIntersection";
+    static constexpr auto name = "s2RectAdd";
 
     static FunctionPtr create(ContextPtr)
     {
-        return std::make_shared<FunctionS2RectIntersection>();
+        return std::make_shared<FunctionS2RectAdd>();
     }
 
     std::string getName() const override
@@ -50,15 +50,14 @@ public:
 
     DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override
     {
-        for (size_t i = 0; i < getNumberOfArguments(); ++i)
+        for (size_t index = 0; index < getNumberOfArguments(); ++index)
         {
-            const auto * arg = arguments[i].get();
+            const auto * arg = arguments[index].get();
             if (!WhichDataType(arg).isUInt64()) {
                 throw Exception(
                     ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT,
                     "Illegal type {} of argument {} of function {}. Must be UInt64",
-                    arg->getName(), i, getName()
-                    );
+                    arg->getName(), index, getName());
             }
         }
 
@@ -69,39 +68,31 @@ public:
 
     ColumnPtr executeImpl(const ColumnsWithTypeAndName & arguments, const DataTypePtr &, size_t input_rows_count) const override
     {
-        const auto * col_lo1 = arguments[0].column.get();
-        const auto * col_hi1 = arguments[1].column.get();
-        const auto * col_lo2 = arguments[2].column.get();
-        const auto * col_hi2 = arguments[3].column.get();
+        const auto * col_lo = arguments[0].column.get();
+        const auto * col_hi = arguments[1].column.get();
+        const auto * col_point = arguments[2].column.get();
 
         auto col_res_first = ColumnUInt64::create();
         auto col_res_second = ColumnUInt64::create();
 
         auto & vec_res_first = col_res_first->getData();
-        vec_res_first.resize(input_rows_count);
+        vec_res_first.reserve(input_rows_count);
 
         auto & vec_res_second = col_res_second->getData();
-        vec_res_second.resize(input_rows_count);
+        vec_res_second.reserve(input_rows_count);
 
         for (const auto row : collections::range(0, input_rows_count))
         {
-            const UInt64 lo1 = col_lo1->getUInt(row);
-            const UInt64 hi1 = col_hi1->getUInt(row);
-            const UInt64 lo2 = col_lo2->getUInt(row);
-            const UInt64 hi2 = col_hi2->getUInt(row);
+            const UInt64 lo = col_lo->getUInt(row);
+            const UInt64 hi = col_hi->getUInt(row);
+            const UInt64 point = col_point->getUInt(row);
 
-            S2CellId id_lo1(lo1);
-            S2CellId id_hi1(hi1);
-            S2CellId id_lo2(lo2);
-            S2CellId id_hi2(hi2);
+            S2LatLngRect rect(S2CellId(lo).ToLatLng(), S2CellId(hi).ToLatLng());
 
-            S2LatLngRect rect1(id_lo1.ToLatLng(), id_hi1.ToLatLng());
-            S2LatLngRect rect2(id_lo2.ToLatLng(), id_hi2.ToLatLng());
+            rect.AddPoint(S2CellId(point).ToPoint());
 
-            S2LatLngRect rect_intersection = rect1.Intersection(rect2);
-
-            vec_res_first[row] = S2CellId(rect_intersection.lo()).id();
-            vec_res_second[row] = S2CellId(rect_intersection.hi()).id();
+            vec_res_first.emplace_back(S2CellId(rect.lo()).id());
+            vec_res_second.emplace_back(S2CellId(rect.hi()).id());
         }
 
         return ColumnTuple::create(Columns{std::move(col_res_first), std::move(col_res_second)});
@@ -111,9 +102,9 @@ public:
 
 }
 
-void registerFunctionS2RectIntersection(FunctionFactory & factory)
+void registerFunctionS2RectAdd(FunctionFactory & factory)
 {
-    factory.registerFunction<FunctionS2RectIntersection>();
+    factory.registerFunction<FunctionS2RectAdd>();
 }
