+++ src/Functions/s2RectContains.cpp
@@ -5,7 +5,9 @@
 #if USE_S2_GEOMETRY
 
 #include <Columns/ColumnsNumber.h>
+#include <Columns/ColumnTuple.h>
 #include <DataTypes/DataTypesNumber.h>
+#include <DataTypes/DataTypeTuple.h>
 #include <Functions/FunctionFactory.h>
 #include <Common/typeid_cast.h>
 #include <common/range.h>
@@ -26,15 +28,14 @@ namespace ErrorCodes
 namespace
 {
 
-/// TODO: Comment this
-class FunctionRadiansToS2 : public IFunction
+class FunctionS2RectContains : public IFunction
 {
 public:
-    static constexpr auto name = "radiansToS2";
+    static constexpr auto name = "s2RectContains";
 
     static FunctionPtr create(ContextPtr)
     {
-        return std::make_shared<FunctionRadiansToS2>();
+        return std::make_shared<FunctionS2RectContains>();
     }
 
     std::string getName() const override
@@ -42,7 +43,7 @@ public:
         return name;
     }
 
-    size_t getNumberOfArguments() const override { return 2; }
+    size_t getNumberOfArguments() const override { return 4; }
 
     bool useDefaultImplementationForConstants() const override { return true; }
 
@@ -51,35 +52,40 @@ public:
         for (size_t i = 0; i < getNumberOfArguments(); ++i)
         {
             const auto * arg = arguments[i].get();
-            if (!WhichDataType(arg).isFloat64()) {
+            if (!WhichDataType(arg).isUInt64()) {
                 throw Exception(
                     ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT,
-                    "Illegal type {} of argument {} of function {}. Must be Float64",
-                    arg->getName(), i, getName()
-                    );
+                    "Illegal type {} of argument {} of function {}. Must be UInt64",
+                    arg->getName(), i, getName());
             }
         }
 
-        return std::make_shared<DataTypeUInt64>();
+        return std::make_shared<DataTypeUInt8>();
     }
 
     ColumnPtr executeImpl(const ColumnsWithTypeAndName & arguments, const DataTypePtr &, size_t input_rows_count) const override
     {
-        const auto * col_lon = arguments[0].column.get();
-        const auto * col_lat = arguments[1].column.get();
+        const auto * col_lo = arguments[0].column.get();
+        const auto * col_hi = arguments[1].column.get();
+        const auto * col_point = arguments[2].column.get();
 
-        auto dst = ColumnVector<UInt64>::create();
+        auto dst = ColumnVector<UInt8>::create();
         auto & dst_data = dst->getData();
         dst_data.resize(input_rows_count);
 
         for (const auto row : collections::range(0, input_rows_count))
         {
-            const double lon = col_lon->getFloat64(row);
-            const double lat = col_lat->getFloat64(row);
+            const UInt64 lo = col_lo->getUInt(row);
+            const UInt64 hi = col_hi->getUInt(row);
+            const UInt64 point = col_point->getUInt(row);
 
-            S2CellId id(S2LatLng::FromRadians(lat, lon));
+            S2CellId id_lo(lo);
+            S2CellId id_hi(hi);
+            S2CellId id_point(point);
 
-            dst_data[row] = id.id();
+            S2LatLngRect rect(id_lo.ToLatLng(), id_hi.ToLatLng());
+
+            dst_data[row] = rect.Contains(id_point.ToLatLng());
         }
 
         return dst;
@@ -89,9 +95,9 @@ public:
 
 }
 
-void registerFunctionRadiansToS2(FunctionFactory & factory)
+void registerFunctionS2RectContains(FunctionFactory & factory)
 {
-    factory.registerFunction<FunctionRadiansToS2>();
+    factory.registerFunction<FunctionS2RectContains>();
 }
