+++ src/Functions/s2ToGeo.cpp
@@ -4,15 +4,17 @@
 
 #if USE_S2_GEOMETRY
 
-#include <Columns/ColumnArray.h>
 #include <Columns/ColumnsNumber.h>
+#include <Columns/ColumnTuple.h>
 #include <DataTypes/DataTypesNumber.h>
-#include <DataTypes/DataTypeArray.h>
+#include <DataTypes/DataTypeTuple.h>
 #include <Functions/FunctionFactory.h>
 #include <Common/typeid_cast.h>
 #include <common/range.h>
 
 #include "s2_fwd.h"
+
+class S2CellId;
 
 namespace DB
 {
@@ -26,15 +28,17 @@ namespace ErrorCodes
 namespace
 {
 
-/// TODO: Comment this
-class FunctionS2GetNeighbors : public IFunction
+/**
+ *  Returns a point (longitude, latitude) in degrees
+ */
+class FunctionS2ToGeo : public IFunction
 {
 public:
-    static constexpr auto name = "S2GetNeighbors";
+    static constexpr auto name = "s2ToGeo";
 
     static FunctionPtr create(ContextPtr)
     {
-        return std::make_shared<FunctionS2GetNeighbors>();
+        return std::make_shared<FunctionS2ToGeo>();
     }
 
     std::string getName() const override
@@ -54,50 +58,48 @@ public:
             throw Exception(
                 ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT,
                 "Illegal type {} of argument {} of function {}. Must be Float64",
-                arg->getName(), 1, getName()
-                );
+                arg->getName(), 1, getName());
         }
 
-        return std::make_shared<DataTypeArray>(std::make_shared<DataTypeUInt64>());
+        DataTypePtr element = std::make_shared<DataTypeFloat64>();
+
+        return std::make_shared<DataTypeTuple>(DataTypes{element, element});
     }
 
     ColumnPtr executeImpl(const ColumnsWithTypeAndName & arguments, const DataTypePtr &, size_t input_rows_count) const override
     {
         const auto * col_id = arguments[0].column.get();
 
-        auto dst = ColumnArray::create(ColumnUInt64::create());
-        auto & dst_data = dst->getData();
-        auto & dst_offsets = dst->getOffsets();
-        dst_offsets.resize(input_rows_count);
-        auto current_offset = 0;
+        auto col_longitude = ColumnFloat64::create();
+        auto col_latitude = ColumnFloat64::create();
+
+        auto & longitude = col_longitude->getData();
+        longitude.reserve(input_rows_count);
+
+        auto & latitude = col_latitude->getData();
+        latitude.reserve(input_rows_count);
 
         for (const auto row : collections::range(0, input_rows_count))
         {
             const UInt64 id = col_id->getUInt(row);
 
-            S2CellId cell_id(id);
-            S2CellId neighbors[4];
-            cell_id.GetEdgeNeighbors(neighbors);
+            S2Point point = S2CellId(id).ToPoint();
+            S2LatLng ll(point);
 
-            dst_data.reserve(dst_data.size() + 4);
-            for (auto & neighbor : neighbors)
-            {
-                ++current_offset;
-                dst_data.insert(neighbor.id());
-            }
-            dst_offsets[row] = current_offset;
+            longitude.emplace_back(ll.lng().degrees());
+            latitude.emplace_back(ll.lat().degrees());
         }
 
-        return dst;
+        return ColumnTuple::create(Columns{std::move(col_longitude), std::move(col_latitude)});
     }
 
 };
 
 }
 
-void registerFunctionS2GetNeighbors(FunctionFactory & factory)
+void registerFunctionS2ToGeo(FunctionFactory & factory)
 {
-    factory.registerFunction<FunctionS2GetNeighbors>();
+    factory.registerFunction<FunctionS2ToGeo>();
 }
