--- tests/queries/0_stateless/01849_degrees_to_s2.reference
@@ -1,4 +1,0 @@
-4573520603753569857
-4573517609713934091
-4035225266123964415
-936559659395828129
