--- tests/queries/0_stateless/01849_degrees_to_s2.sql
@@ -1,4 +1,0 @@
-select degreesToS2(55.77922738, 37.63098076);
-select degreesToS2(55.76324100, 37.66018300);
-select degreesToS2(90.0, 45.0);
-select degreesToS2(-12.0, 34.0);
