+++ tests/queries/0_stateless/01849_geoToS2.reference
@@ -1,0 +1,43 @@
+Checking s2 index generation.
+(-19.82614013111778,-41.19291183249827)	145638248314527735
+(-41.56412828256075,-16.777072680829264)	525948609053544771
+(40.76827373895363,2.853562616147134)	1723620528513494869
+(8.774109215249668,-3.271373776817451)	1913723177026859677
+(54.7243525263686,-19.21060843697615)	2414200527355011557
+(49.942295220850404,-18.63385558246411)	2446780491370050277
+(125.93850250627281,25.519361668875952)	3814912406305146887
+(51.39037388040172,29.368252361755857)	4590287096029015617
+(-165.90797307310456,54.20517787077579)	6317132534461540395
+(140.4288338857567,28.399754752831992)	6944470717484584123
+(170.14574840189854,7.606448822713084)	7280210779810726069
+(-170.42364912433663,-10.102188288980733)	8094352344009072653
+(-168.25897915006252,-38.27117023780382)	8295275405228382549
+(-46.71824230901231,5.161978621886426)	13251733624047612511
+(-64.36499761086276,-13.206225582160274)	10654167528317614093
+(-61.76193800786795,-24.969589107565216)	10670400906708524493
+(-79.24545956192031,-22.940848730236024)	10868726821406045765
+(74.00610377406458,-68.32123992734591)	12793606480989360605
+(10.610774069458158,-64.18410328814072)	13202270384266773975
+(-89.81096210929424,-57.01398354986957)	13606307743304496003
+(-19.82614,-41.19291)	(-19.82614,-41.19291)	ok
+(-41.56413,-16.77707)	(-41.56413,-16.77707)	ok
+(40.76827,2.85356)	(40.76827,2.85356)	ok
+(8.77411,-3.27137)	(8.77411,-3.27137)	ok
+(54.72435,-19.21061)	(54.72435,-19.21061)	ok
+(49.94229,-18.63386)	(49.94229,-18.63386)	ok
+(125.9385,25.51936)	(125.9385,25.51936)	ok
+(51.39037,29.36825)	(51.39037,29.36825)	ok
+(-165.90797,54.20518)	(-165.90797,54.20518)	ok
+(140.42883,28.39976)	(140.42883,28.39976)	ok
+(170.14575,7.60645)	(170.14575,7.60645)	ok
+(-170.42365,-10.10219)	(-170.42365,-10.10219)	ok
+(-168.25898,-38.27117)	(-168.25898,-38.27117)	ok
+(5.16198,-46.71824)	(5.16198,-46.71824)	ok
+(-64.365,-13.20623)	(-64.365,-13.20623)	ok
+(-61.76194,-24.96959)	(-61.76194,-24.96959)	ok
+(-79.24546,-22.94085)	(-79.24546,-22.94085)	ok
+(74.0061,-68.32124)	(74.0061,-68.32124)	ok
+(10.61077,-64.1841)	(10.61077,-64.1841)	ok
+(-89.81096,-57.01398)	(-89.81096,-57.01398)	ok
+(-45,-35.264389716294055)
+4864204703484167331
