+++ tests/queries/0_stateless/01853_s2_cells_intersect.sql
@@ -1,2 +1,2 @@
-select S2CellsIntersect(9926595209846587392, 9926594385212866560);
-select S2CellsIntersect(9926595209846587392, 9937259648002293760);
+select s2CellsIntersect(9926595209846587392, 9926594385212866560);
+select s2CellsIntersect(9926595209846587392, 9937259648002293760);
