+++ tests/queries/0_stateless/01854_s2_cap_contains.reference
@@ -1,2 +1,5 @@
 1
 0
+1
+0
+1
