+++ tests/queries/0_stateless/01854_s2_cap_contains.sql
@@ -1,2 +1,11 @@
-select S2CapContains(1157339245694594829, 1.0, 1157347770437378819);
-select S2CapContains(1157339245694594829, 1.0, 1152921504606846977);
+select s2CapContains(1157339245694594829, 1.0, 1157347770437378819);
+select s2CapContains(1157339245694594829, 1.0, 1152921504606846977);
+select s2CapContains(1157339245694594829, 3.14, 1157339245694594829);
+
+select s2CapContains(nan, 3.14, 1157339245694594829); -- { serverError 43 }
+select s2CapContains(1157339245694594829, nan, 1157339245694594829); -- { serverError 43 }
+select s2CapContains(1157339245694594829, 3.14, nan); -- { serverError 43 }
+
+
+select s2CapContains(toUInt64(-1), -1.0, toUInt64(-1));
+select s2CapContains(toUInt64(-1), 9999.9999, toUInt64(-1));
