+++ tests/queries/0_stateless/01854_s2_cap_union.reference
@@ -1,0 +1,3 @@
+(4534655147792050737,60.2088283994957)
+(1157339245694594829,-57.29577951308232)
+(1157339245694594829,180)
